+++ src/components/Confirm/Actions.js
@@ -1,16 +1,20 @@
 import { h } from 'preact'
+import { Button } from '@onfido/castor-react'
 import classNames from 'classnames'
-import Button from '../Button'
 import { localised } from '../../locales'
+import theme from '../Theme/style.scss'
 import style from './style.scss'
 
 const RetakeAction = localised(({ retakeAction, translate, singleAction }) => (
   <Button
     onClick={retakeAction}
-    className={singleAction ? null : style.retakeAction}
-    variants={
-      singleAction ? ['primary', 'lg', 'centered'] : ['secondary', 'sm']
+    variant={singleAction ? 'primary' : 'secondary'}
+    className={
+      singleAction
+        ? classNames(theme['button-lg'], theme['button-centered'])
+        : classNames(theme['button-sm'], style.retakeAction)
     }
+    data-onfido-qa="redo-action-btn"
   >
     {translate(
       singleAction
@@ -23,9 +27,11 @@ const RetakeAction = localised(({ retakeAction, translate, singleAction }) => (
 const ConfirmAction = localised(
   ({ confirmAction, isUploading, translate, error }) => (
     <Button
-      variants={['primary', 'sm']}
+      variant="primary"
+      className={theme['button-sm']}
       onClick={confirmAction}
       disabled={isUploading}
+      data-onfido-qa="confirm-action-btn"
     >
       {error.type === 'warn'
         ? translate('doc_confirmation.button_primary_upload_anyway')