+++ src/components/ProofOfAddress/Guidance/index.js
@@ -1,6 +1,7 @@
 import { h } from 'preact'
+import { Button } from '@onfido/castor-react'
+import classNames from 'classnames'
 import PageTitle from '../../PageTitle'
-import Button from '../../Button'
 import { trackComponent } from '../../../Tracker'
 import { localised } from '../../../locales'
 import { POA_GUIDANCE_LOCALES_MAPPING } from '~utils/localesMapping'
@@ -42,7 +43,12 @@ const Guidance = ({
         </div>
       </div>
       <div className={theme.contentMargin}>
-        <Button variants={['primary', 'centered', 'lg']} onClick={nextStep}>
+        <Button
+          variant="primary"
+          className={classNames(theme['button-centered'], theme['button-lg'])}
+          onClick={nextStep}
+          data-onfido-qa="poa-continue-btn"
+        >
           {translate('poa_guidance.button_primary')}
         </Button>
       </div>