+++ src/components/Uploader/index.js
@@ -1,4 +1,5 @@
 import { h, Component } from 'preact'
+import { Button } from '@onfido/castor-react'
 import classNames from 'classnames'
 import { isDesktop } from '~utils'
 import { camelCase } from '~utils/string'
@@ -7,7 +8,6 @@ import { trackComponentAndMode } from '../../Tracker'
 import { localised } from '../../locales'
 import CustomFileInput from '../CustomFileInput'
 import PageTitle from '../PageTitle'
-import Button from '../Button'
 import UploadError from './Error'
 import theme from '../Theme/style.scss'
 import style from './style.scss'
@@ -32,7 +32,12 @@ const MobileUploadArea = ({
         capture
       >
         <Button
-          variants={isPoA ? ['secondary', 'sm'] : ['centered', 'primary', 'lg']}
+          variant={isPoA ? 'secondary' : 'primary'}
+          className={
+            isPoA
+              ? theme['button-sm']
+              : classNames(theme['button-centered'], theme['button-lg'])
+          }
           disabled={isUploading}
         >
           {translate('photo_upload.button_take_photo')}
@@ -40,7 +45,11 @@ const MobileUploadArea = ({
       </CustomFileInput>
       {isPoA && (
         <CustomFileInput onChange={onFileSelected}>
-          <Button variants={['primary', 'sm']} disabled={isUploading}>
+          <Button
+            variant="primary"
+            className={theme['button-sm']}
+            disabled={isUploading}
+          >
             {translate(
               isDesktop
                 ? 'doc_submit.button_link_upload'
@@ -63,7 +72,8 @@ const PassportMobileUploadArea = ({
     {children}
     <div className={style.buttons}>
       <Button
-        variants={['centered', 'primary', 'lg']}
+        variant="primary"
+        className={classNames(theme['button-centered'], theme['button-lg'])}
         disabled={isUploading}
         onClick={nextStep}
       >
@@ -94,8 +104,12 @@ const DesktopUploadArea = ({
     <div>
       {!mobileFlow && ( // Hide for mobileFlow on desktop browser as `test` Node environment has restrictedXDevice set to false
         <Button
-          variants={['centered', 'primary', 'lg']}
-          className={style.crossDeviceButton}
+          variant="primary"
+          className={classNames(
+            theme['button-centered'],
+            theme['button-lg'],
+            style.crossDeviceButton
+          )}
           onClick={() => changeFlowTo('crossDeviceSteps')}
           disabled={isUploading}
         >