+++ src/components/Welcome/index.js
@@ -1,7 +1,8 @@
 import { h } from 'preact'
 import ScreenLayout from '../Theme/ScreenLayout'
+import { Button } from '@onfido/castor-react'
+import classNames from 'classnames'
 import PageTitle from '../PageTitle'
-import Button from '../Button'
 import { trackComponent } from '../../Tracker'
 import { localised } from '../../locales'
 import { buildIteratorKey } from '~utils'
@@ -38,7 +39,12 @@ const WelcomeActions = ({ nextButton, nextStep, translate }) => {
 
   return (
     <div className={theme.contentMargin}>
-      <Button onClick={nextStep} variants={['centered', 'primary', 'lg']}>
+      <Button
+        variant="primary"
+        className={classNames(theme['button-centered'], theme['button-lg'])}
+        onClick={nextStep}
+        data-onfido-qa="welcome-next-btn"
+      >
         {welcomeNextButton}
       </Button>
     </div>