+++ src/components/crossDevice/CrossDeviceLink/index.js
@@ -1,9 +1,9 @@
 import { h, Component } from 'preact'
+import { Button } from '@onfido/castor-react'
 import classNames from 'classnames'
 import { performHttpReq } from '~utils/http'
 import { formatError } from '~utils/onfidoApi'
 import Spinner from '../../Spinner'
-import Button from '../../Button'
 import CopyLink from './CopyLink'
 import PhoneNumberInputLazy from '../../PhoneNumberInput/Lazy'
 import QRCodeGenerator from '../../QRCode'
@@ -261,13 +261,14 @@ class CrossDeviceLinkUI extends Component {
               />
             </div>
             <Button
-              ariaLive="polite"
-              ariaRelevant="text"
-              ariaBusy={sending}
+              variant="primary"
               className={classNames(style.btn, { [style.sending]: sending })}
-              variants={['primary']}
               onClick={this.handleSendSmsLinkClick}
               disabled={sending || invalidNumber}
+              aria-busy={sending}
+              aria-live="polite"
+              aria-relevant="text"
+              data-onfido-qa="cross-device-send-link-btn"
             >
               {translate(buttonCopyKey)}
             </Button>